increase up-time counters to unsigned 64bit integer
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@2420 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ simcraft.h
@@ -1624,7 +1624,7 @@ struct proc_t
 struct uptime_t
 {
   std::string name_str;
-  int up, down;
+  uint64_t up, down;
   uptime_t* next;
   uptime_t( const std::string& n ) : name_str( n ), up( 0 ), down( 0 ) {}
   void   update( bool is_up ) { if ( is_up ) up++; else down++; }
